Refactor import paths for RestClient in RTC client files to maintain consistency
